Separate every command into its own file
