+++ JPDB Bot/Commands/ChangeLog.cs
@@ -1,0 +1,48 @@
+using DSharpPlus.CommandsNext;
+using DSharpPlus.CommandsNext.Attributes;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+using HtmlAgilityPack;
+using DiscordBot;
+
+namespace JPDB_Bot.Commands
+{
+    public class ChangeLog : BaseCommandModule
+    {
+        [Command("changelog")]
+        [Cooldown(1, 20, CooldownBucketType.User)]
+        [Description("Shows the latest addition to the changelog")]
+        public async Task changeLog(CommandContext ctx)
+        {
+            Program.PrintCommandUse(ctx.User.Username, ctx.Message.Content);
+            string Date = String.Empty;
+
+            /*WebClient Client = new WebClient();
+            Client.Encoding = System.Text.Encoding.UTF8;
+            string HTML = "";
+            string sniptemp = string.Empty;
+            HTML = Client.DownloadString(new Uri("https://jpdb.io/changelog"));
+            int snipIndex = HTML.IndexOf("<h5 id=") + 8;
+            HTML = HTML.Substring(snipIndex);
+            snipIndex = HTML.IndexOf(">") + 1;
+            HTML = HTML.Substring(snipIndex);
+
+            sniptemp = HTML;
+            snipIndex = sniptemp.IndexOf("<");
+            sniptemp = sniptemp.Substring(0, snipIndex);
+
+            var changelogEmbed = new discordembed*/
+
+            HtmlWeb web = new HtmlWeb();
+            var htmlDoc = web.Load("https://jpdb.io/changelog");
+            string ChangeDate = htmlDoc.DocumentNode.SelectNodes("/html/body/div[2]/h5[1]")
+                .First()
+                .InnerHtml;
+            string Information = htmlDoc.DocumentNode.SelectNodes("/html/body/div[2]/h5[1]").First().NextSibling
+                .InnerHtml;
+
+            await ctx.RespondAsync(ChangeDate).ConfigureAwait(false);
+        }
+    }
+}
