+++ JPDB Bot/Commands/Content.cs
@@ -1,0 +1,91 @@
+using DSharpPlus.CommandsNext;
+using DSharpPlus.CommandsNext.Attributes;
+using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Threading.Tasks;
+using DiscordBot;
+
+namespace JPDB_Bot.Commands
+{
+    public class Content : BaseCommandModule
+    {
+        [Command("content")]
+        [Cooldown(3, 10, CooldownBucketType.User)]
+        [Description("Search for content in the JPDB database\nFor example: !content \"steins gate\"")]
+        public async Task SearchContent(CommandContext ctx,
+            [DescriptionAttribute("Name of the content you are searching")]
+            string searchString)
+        {
+            Program.PrintCommandUse(ctx.User.Username, ctx.Message.Content);
+            //await ctx.Channel.SendMessageAsync("Searching for " + searchString + "...").ConfigureAwait(false);
+
+
+            WebClient Client = new WebClient();
+            Client.Encoding = System.Text.Encoding.UTF8;
+            string HTML = "";
+            int snipIndex = -1;
+
+            string OriginalFilter = string.Empty;
+            OriginalFilter = "anime";
+
+            if (HTML.Length > 250) return;
+
+            string wordTemp = "";
+            string URL = "";
+            List<string> wordIDs = new List<string>() { };
+            URL = "https://jpdb.io/prebuilt_decks?q=" + searchString;
+            try
+            {
+                HTML = Client.DownloadString(new Uri(URL));
+            }
+            catch (Exception ex)
+            {
+                Program.PrintError(ex.Message);
+                return;
+            }
+
+            snipIndex = HTML.IndexOf("30rem;\">") + 8;
+
+            if (snipIndex == 7)
+            {
+                await ctx.RespondAsync("No content found UwU").ConfigureAwait(false);
+                return;
+            }
+
+            wordTemp = HTML.Substring(snipIndex);
+            HTML = wordTemp;
+
+            snipIndex = wordTemp.IndexOf("<");
+            HTML = wordTemp.Substring(snipIndex);
+            string contentName = wordTemp.Substring(0, snipIndex);
+
+            snipIndex = HTML.IndexOf("margin-top: 0.5rem;\">") + 22;
+            wordTemp = HTML.Substring(snipIndex);
+            snipIndex = wordTemp.IndexOf("/") + 1;
+            wordTemp = wordTemp.Substring(snipIndex);
+            snipIndex = wordTemp.IndexOf("\"");
+            wordTemp = "https://jpdb.io/" + wordTemp.Substring(0, snipIndex);
+
+            await ctx.RespondAsync("Found " + contentName + ":\n" + wordTemp).ConfigureAwait(false);
+
+            int Frequency = 1;
+            if (wordTemp.Contains(">") == false && wordTemp.Contains("<") == false & wordTemp.Contains("=") == false &
+                wordTemp.Contains("-") == false)
+            {
+                try
+                {
+                    for (var I = 1; I <= Frequency; I++)
+                        wordIDs.Add(wordTemp);
+                }
+                catch (Exception ex)
+                {
+                    Program.PrintError(ex.Message);
+                    return;
+                }
+            }
+
+            snipIndex = HTML.IndexOf("#a") + 3;
+        }
+    }
+}
